Add forged-session Playwright specs
Add Playwright specs that exercise OS with directly minted Iterate session JWTs and seeded project fixtures.

Consolidate local OS dev server discovery/startup into apps/os/scripts/dev.ts so specs can reuse sticky per-worktree servers or target preview deployments through OS_BASE_URL.

Expose explicit --port and --keep-alive options for Playwright webServer, remove spec-local dev/auth shims, and document the Doppler env contract for forged sessions.
